pyproject.toml
[tool.poetry]
name = "resume-update"
version = "0.1.0"
description = ""
authors = ["arturkalimullin <noti1996@gmail.com>"]
readme = "README.md"
packages = [{include = "resume_update"}]

[tool.poetry.dependencies]
python = "^3.11"
schedule = "^1.1.0"
beautifulsoup4 = "^4.12.0"
requests = "^2.28.2"
selenium = "^4.8.3"
python-dotenv = "^1.0.1"
selenium-stealth = "^1.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
